requirements.txt
yt-dlp>=2023.10.7
Pillow>=11.0.0
requests>=2.32.0